add a few justwatch providers
